reworked config
