+++ examples/notebooks/metaScript.bash
