+++ isd_s3_config.ini
@@ -1,0 +1,17 @@
+[common]
+base_path = /glade/u/home/rdadata
+
+[default]
+s3_url = https://stratus.ucar.edu
+credentials = /glade/u/home/rdadata/.aws/credentials
+bucket = rda-data
+
+[logging]
+logpath = ${common:base_path}/dssdb/log
+logfile = isd-s3.log
+dbgfile = isd-s3.dbg
+loglevel = info 
+maxbytes = 2000000
+backupcount = 1
+logfmt = %(asctime)s - %(name)s - %(funcName)s - %(levelname)s - %(message)s
+dbgfmt = %(asctime)s - %(name)s - %(funcName)s - %(lineno)d - %(levelname)s - %(message)s
